+++ src/Farmer/Arm/Network.fs
@@ -45,6 +45,7 @@ type SubnetDelegation =
 type Subnet =
     { Name : ResourceName
       Prefix : string
+      NetworkSecurityGroup : LinkedResource option
       Delegations : SubnetDelegation list
       ServiceEndpoints : (Network.EndpointServiceType * Location list) list
       AssociatedServiceEndpointPolicies : ResourceId list 
@@ -60,7 +61,14 @@ type VirtualNetwork =
     interface IArmResource with
         member this.ResourceId = virtualNetworks.resourceId this.Name
         member this.JsonModel =
-            {| virtualNetworks.Create(this.Name, this.Location, tags = this.Tags) with
+            let dependencies =
+                seq {
+                    for subnet in this.Subnets do
+                        match subnet.NetworkSecurityGroup with
+                        | Some (Managed id) -> id
+                        | _ -> ()
+                } |> Set
+            {| virtualNetworks.Create(this.Name, this.Location, dependsOn=dependencies, tags = this.Tags) with
                 properties =
                     {| addressSpace = {| addressPrefixes = this.AddressSpacePrefixes |}
                        subnets =
@@ -69,6 +77,10 @@ type VirtualNetwork =
                             {| name = subnet.Name.Value
                                properties =
                                 {| addressPrefix = subnet.Prefix
+                                   networkSecurityGroup = 
+                                       subnet.NetworkSecurityGroup
+                                       |> Option.map(fun nsg -> {| id = nsg.ResourceId.ArmExpression.Eval() |}) 
+                                       |> Option.defaultValue Unchecked.defaultof<_> 
                                    delegations =
                                        subnet.Delegations
                                        |> List.map (fun delegation ->